requirements.txt
# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0
scikit-image>=0.21.0

# Image processing and augmentation
albumentations>=1.3.1
imagehash>=4.3.1

# Data handling
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Experiment tracking and ML utilities
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Model serving and APIs (optional)
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

# OpenAI API
openai>=1.0.0
